Fix minor type error.
Monotone-Parent: 7b11ea4cf2e35b6af66c11df6cd6a2e16a753e7e
Monotone-Revision: 91fd0c4ac4360430cc6332ee6e1d195013b2bfc1

+++ include/simple_cfd/numeric/cast.hpp
@@ -19,7 +19,7 @@ struct RawConverter
 template<>
 struct RawConverter<cln::cl_R, float>
 {
-  static double low_level_convert(const cln::cl_R& s) 
+  static float low_level_convert(const cln::cl_R& s) 
   { 
     return cln::float_approx(s);
   }
